Animate zandi properly for cleft visitors
Because of the way NPC Spawner Mods work, Zandi would only be properly
animated on the first user to join the Cleft instance. Others would not
see the NPC Spawner Mod notification, and as such, would have a NoneType
in place of the reference to the Zandi avatar, which would cause errors.