Make Zandi's Multistage Behavior player-local
There are race conditions that occur when arriving in the Cleft while
the MultiStage is running a behavior, which can cause zandi to move
to a standing position when Zandi finally finishes its behavior. To
avoid those, we simply make the MultiStage Behavior Mod local to the
player by disabling netForce and netProp. They will still play in
sync since it is activated by other events that are themselves
netPropagated.